#sorting/move_functions_utils_folder: functions moved to utils folder and path updated in build.js
