+++ portfolio.js
@@ -1,7 +1,7 @@
 const fs = require("fs-extra");
 const readline = require("readline");
 const path = require("path");
-const { build } = require("./build.js");
+const { build } = require("./utils/build.js");
 
 const rl = readline.createInterface({
   input: process.stdin,
@@ -52,10 +52,10 @@ async function setConfig() {
   let cfg = {};
 
   try {
-    const exists = await fs.existsSync(path.join(__dirname, "config.json"));
+    const exists = await fs.existsSync(path.join(__dirname, "utils/config.json"));
 
     if (exists) {
-      cfg = await fs.readJson("config.json", { throws: false });
+      cfg = await fs.readJson("utils/config.json", { throws: false });
     }
 
     data = await question("\nGitHub username: ", cfg.username);
@@ -110,9 +110,9 @@ async function setConfig() {
       }
     } while (choice != "no");
 
-    await fs.writeJson("./config.json", cfg, { spaces: "\t" });
-    const obj = await fs.readJson("./config.json", { throws: false });
-    console.log("\nConfig written to ./config.json");
+    await fs.writeJson("./utils/config.json", cfg, { spaces: "\t" });
+    const obj = await fs.readJson("./utils/config.json", { throws: false });
+    console.log("\nConfig written to ./utils/config.json");
     console.log(obj); // => null
 
     console.log("🎉 Config set!");
